Move and clean up timeframe utils
